v0.1.0 GameApp.Game

GameApp.Game defines a struct that encapsulates Game state as well as many functions that advance the game state based on actions that players can take.

Types

game()

game() :: %GameApp.Game{
  creator: term(),
  funmaster: term(),
  funmaster_order: term(),
  phase: term(),
  players: term(),
  round_number: term(),
  rounds: term(),
  scores: term(),
  shortcode: term(),
  winner: term()
}

game_state()

game_state() ::
  :lobby
  | :game_start
  | :round_start
  | :prompt_selection
  | :reaction_selection
  | :winner_selection
  | :round_end
  | :game_end

player()

player() :: %{id: String.t()}

prompt()

prompt() :: String.t()

reaction()

reaction() :: String.t()

round()

round() :: %GameApp.Round{number: term(), reactions: term(), winner: term()}

shortcode()

shortcode() :: String.t()

Functions

create(shortcode, creator)

create(shortcode(), player()) :: game()

Creates a game.

Returns %Game{}.

Examples

iex> GameApp.Game.create("ABCD", %{id: "Gamer"})
%GameApp.Game{
  shortcode: "ABCD",
  creator: %{id: "Gamer"},
  players: %{"Gamer" => %{id: "Gamer"}},
  scores: %{"Gamer" => 0}
}

player_join(game, player)

player_join(game(), player()) :: game()

Adds a player to a game.

Returns %Game{}.

Examples

iex> g = GameApp.Game.create("ABCD", %{id: "Gamer"})
iex> GameApp.Game.player_join(g, %{id: "Gamer2"})
%GameApp.Game{
  shortcode: "ABCD",
  creator: %{id: "Gamer"},
  players: %{"Gamer" => %{id: "Gamer"}, "Gamer2" => %{id: "Gamer2"}},
  scores: %{"Gamer" => 0, "Gamer2" => 0}
}

player_leave(game, player)

player_leave(game(), player()) :: game()

Removes a player from a game.

Returns %Game{}.

Examples

iex> g = GameApp.Game.create("ABCD", %{id: "Gamer"})
iex> g = GameApp.Game.player_join(g, %{id: "Gamer2"})
iex> GameApp.Game.player_leave(g, %{id: "Gamer2"})
%GameApp.Game{
  shortcode: "ABCD",
  creator: %{id: "Gamer"},
  players: %{"Gamer" => %{id: "Gamer"}},
  scores: %{"Gamer" => 0, "Gamer2" => 0}
}

select_prompt(game, prompt)

select_prompt(game(), prompt()) :: game()

select_reaction(game, player, reaction)

select_reaction(game(), player(), reaction()) :: game()

select_round_winner(game, player)

select_round_winner(game(), player()) :: game()

start_game(game)

start_game(game()) :: game()

Starts a game.

Returns %Game{}.

Examples

iex> g = GameApp.Game.create("ABCD", %{id: "Gamer"})
iex> GameApp.Game.start_game(g)
%GameApp.Game{
  shortcode: "ABCD",
  phase: :game_start,
  creator: %{id: "Gamer"},
  players: %{"Gamer" => %{id: "Gamer"}},
  scores: %{"Gamer" => 0}
}

start_prompt_selection(game)

start_prompt_selection(game()) :: game()

start_round(game)

start_round(game()) :: game()

start_round_end(game)

start_round_end(game()) :: game()