v0.1.0 GameApp.Round

GameApp.Round defines a struct that encapsulates Round state as well as functions that update the round state.

Types

t()

t() :: %GameApp.Round{
  number: integer(),
  prompt: String.t() | nil,
  reactions: map(),
  winner: GameApp.Player.t() | nil
}

Functions

create(number)

create(integer()) :: GameApp.Round.t()

Creates a round for the given round number.

Examples

iex> GameApp.Round.create(1)
%GameApp.Round{
  number: 1,
  winner: nil,
  reactions: %{}
}

remove_reaction(round, id)

remove_reaction(GameApp.Round.t(), String.t()) :: GameApp.Round.t()

Removes the reaction for a player in a round.

Examples

iex> r = GameApp.Round.create(1)
iex> r = GameApp.Round.set_reaction(r, "Gamer", "OMG!")
iex> GameApp.Round.remove_reaction(r, "Gamer")
%GameApp.Round{
  number: 1,
  winner: nil,
  reactions: %{}
}

set_prompt(round, prompt)

set_prompt(GameApp.Round.t(), String.t()) :: GameApp.Round.t()

Sets the prompt for a round.

Examples

iex> r = GameApp.Round.create(1)
iex> GameApp.Round.set_prompt(r, "Wat")
%GameApp.Round{
  number: 1,
  prompt: "Wat",
  winner: nil,
  reactions: %{}
}

set_reaction(round, id, reaction)

set_reaction(GameApp.Round.t(), String.t(), String.t()) :: GameApp.Round.t()

Sets the reaction for a player in a round.

Examples

iex> r = GameApp.Round.create(1)
iex> GameApp.Round.set_reaction(r, "Gamer", "OMG!")
%GameApp.Round{
  number: 1,
  winner: nil,
  reactions: %{
    "Gamer" => "OMG!"
  }
}

set_winner(round, winner)

set_winner(GameApp.Round.t(), GameApp.Player.t()) :: GameApp.Round.t()

Sets the winner for a round.

Examples

iex> r = GameApp.Round.create(1)
iex> GameApp.Round.set_winner(r, %{id: "Gamer"})
%GameApp.Round{
  number: 1,
  winner: %{id: "Gamer"},
  reactions: %{}
}