v0.1.0 GameApp.Server

GameApp.Server provides a stateful process that maintains an internal Game state and provides a public API for interacting with the game.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

game_pid(shortcode)

Returns the pid of the game server process registered under the given shortcode, or nil if no process is registered.

join(shortcode, player)

leave(shortcode, player)

start_game(shortcode)

start_link(shortcode, player)

start_round(shortcode)

summary(shortcode)

via_tuple(shortcode)

Returns a tuple used to register and lookup a game server process by name.