v0.1.0 GameApp.Game

GameApp.Game defines a struct that encapsulates game state as well as many functions that advance the game state based on actions that players can take.

Types

game_state()

game_state() ::
  :lobby
  | :game_start
  | :round_start
  | :prompt_selection
  | :reaction_selection
  | :winner_selection
  | :round_end
  | :game_end

t()

t() :: %GameApp.Game{
  creator: GameApp.Player.t(),
  funmaster: GameApp.Player.t() | nil,
  funmaster_order: [String.t()],
  phase: game_state(),
  players: %{optional(String.t()) => GameApp.Player.t()},
  round_number: integer() | nil,
  rounds: [GameApp.Round.t()],
  scores: %{optional(String.t()) => integer()},
  shortcode: String.t(),
  winner: GameApp.Player.t() | nil
}

Functions

create(shortcode, creator)

create(String.t(), GameApp.Player.t()) :: GameApp.Game.t()

Creates a game.

Returns %Game{}.

Examples

iex> Game.create("ABCD", Player.create("1", "Gamer"))
%Game{
  shortcode: "ABCD",
  creator: %Player{id: "1", name: "Gamer"},
  players: %{"1" => %Player{id: "1", name: "Gamer"}},
  scores: %{"1" => 0}
}

player_join(game, player)

player_join(GameApp.Game.t(), GameApp.Player.t()) :: GameApp.Game.t()

Adds a player to a game.

Returns %Game{}.

Examples

iex> g = Game.create("ABCD", Player.create("1", "Gamer"))
iex> Game.player_join(g, Player.create("2", "Gamer2"))
%Game{
  shortcode: "ABCD",
  creator: %Player{id: "1", name: "Gamer"},
  players: %{
    "1" => %Player{id: "1", name: "Gamer"},
    "2" => %Player{id: "2", name: "Gamer2"}
  },
  scores: %{"1" => 0, "2" => 0}
}

player_leave(game, player)

player_leave(GameApp.Game.t(), GameApp.Player.t()) :: GameApp.Game.t()

Removes a player from a game.

Returns %Game{}.

Examples

iex> p1 = Player.create("1", "Gamer")
iex> p2 = Player.create("2", "Gamer2")
iex> g = Game.create("ABCD", p1)
iex> g = Game.player_join(g, p2)
iex> Game.player_leave(g, p2)
%Game{
  shortcode: "ABCD",
  creator: %Player{id: "1", name: "Gamer"},
  players: %{"1" => %Player{id: "1", name: "Gamer"}},
  scores: %{"1" => 0, "2" => 0}
}

select_prompt(game, prompt)

select_prompt(GameApp.Game.t(), String.t()) :: GameApp.Game.t()

select_reaction(game, player, reaction)

select_reaction(GameApp.Game.t(), GameApp.Player.t(), String.t()) ::
  GameApp.Game.t()

select_round_winner(game, player)

select_round_winner(GameApp.Game.t(), GameApp.Player.t()) :: GameApp.Game.t()

start_game(game)

start_game(GameApp.Game.t()) :: GameApp.Game.t()

Starts a game.

Returns %Game{}.

Examples

iex> g = Game.create("ABCD", Player.create("1", "Gamer"))
iex> Game.start_game(g)
%Game{
  shortcode: "ABCD",
  phase: :game_start,
  creator: %Player{id: "1", name: "Gamer"},
  players: %{"1" => %Player{id: "1", name: "Gamer"}},
  scores: %{"1" => 0}
}

start_prompt_selection(game)

start_prompt_selection(GameApp.Game.t()) :: GameApp.Game.t()

start_round(game)

start_round(GameApp.Game.t()) :: GameApp.Game.t()

Starts a round.

Returns %Game{}.

Examples

iex> g = Game.create("ABCD", Player.create("1", "Gamer"))
iex> Game.start_game(g)
%Game{
  shortcode: "ABCD",
  phase: :game_start,
  creator: %Player{id: "1", name: "Gamer"},
  players: %{"1" => %Player{id: "1", name: "Gamer"}},
  scores: %{"1" => 0}
}

start_round_end(game)

start_round_end(GameApp.Game.t()) :: GameApp.Game.t()